+++ akshare/__init__.py
@@ -2318,9 +2318,10 @@ amac_manager_cancelled_info # 中国证券投资基金业协会-信息公示-诚
 1.8.99 add: add futures_news_shmet interface
 1.9.1 fix: fix index_value_name_funddb interface
 1.9.2 fix: fix stock_xgsglb_em interface
+1.9.3 fix: fix fx_quote_baidu interface
 """
 
-__version__ = "1.9.2"
+__version__ = "1.9.3"
 __author__ = "AKFamily"
 
 import sys