+++ akshare/fx/fx_quote_baidu.py
@@ -1,12 +1,14 @@
 #!/usr/bin/env python
 # -*- coding:utf-8 -*-
 """
-Date: 2022/10/17 20:12
+Date: 2023/3/5 18:12
 Desc: 百度股市通-外汇-行情榜单
 https://gushitong.baidu.com/top/foreign-common-%E5%B8%B8%E7%94%A8
 """
+import http.client
+import json
+
 import pandas as pd
-import requests
 
 
 def fx_quote_baidu(symbol: str = "人民币") -> pd.DataFrame:
@@ -22,38 +24,36 @@ def fx_quote_baidu(symbol: str = "人民币") -> pd.DataFrame:
         "人民币": "rmb",
         "美元": "dollar",
     }
-    url = "https://finance.pae.baidu.com/api/getforeignrank"
     num = 0
-    params = {
-        'pn': num,
-        'rn': '20',
-        'type': symbol_map[symbol],
-        'finClientType': 'pc'
-    }
     out_df = pd.DataFrame()
     while True:
         try:
-            r = requests.get(url, params=params)
-            data_json = r.json()
-            temp_df = pd.DataFrame(
-                data_json["Result"]
+            conn = http.client.HTTPSConnection("finance.pae.baidu.com")
+            conn.request(
+                "GET", f"/api/getforeignrank?pn={num}&rn=20&type={symbol_map[symbol]}&finClientType=pc"
             )
+            res = conn.getresponse()
+            data = res.read()
+            data_json = json.loads(data.decode("utf-8"))
+            temp_df = pd.DataFrame(data_json["Result"])
             temp_list = []
             for item in temp_df["list"]:
                 temp_list.append(list(pd.DataFrame(item).T.iloc[1, :].values))
-            value_df = pd.DataFrame(temp_list, columns=pd.DataFrame(item).T.iloc[0, :].values)
+            value_df = pd.DataFrame(
+                temp_list, columns=pd.DataFrame(item).T.iloc[0, :].values
+            )
             big_df = pd.concat([temp_df, value_df], axis=1)
-            del big_df['market']
-            del big_df['list']
-            big_df.columns = [
-                '代码', '名称', '最新价', '涨跌额', '涨跌幅'
-            ]
-            big_df['最新价'] = pd.to_numeric(big_df['最新价'])
-            big_df['涨跌额'] = pd.to_numeric(big_df['涨跌额'])
-            big_df['涨跌幅'] = pd.to_numeric(big_df['涨跌幅'].str.strip("%")) / 100
+            del big_df["market"]
+            del big_df["list"]
+            del big_df["status"]
+            del big_df["icon1"]
+            del big_df["icon2"]
+            big_df.columns = ["代码", "名称", "最新价", "涨跌额", "涨跌幅"]
+            big_df["最新价"] = pd.to_numeric(big_df["最新价"])
+            big_df["涨跌额"] = pd.to_numeric(big_df["涨跌额"])
+            big_df["涨跌幅"] = pd.to_numeric(big_df["涨跌幅"].str.strip("%")) / 100
             out_df = pd.concat([out_df, big_df], ignore_index=True)
             num = num + 20
-            params.update({"pn": num})
         except:
             break
     return out_df