+++ docs/changelog.md
@@ -49,6 +49,10 @@
 
 ## 更新说明
 
+1.9.3 fix: fix fx_quote_baidu interface
+
+    1. 修复 fx_quote_baidu 接口
+
 1.9.2 fix: fix stock_xgsglb_em interface
 
     1. 修复 stock_xgsglb_em 接口
@@ -1955,6 +1959,8 @@
     3. 修改 air_quality_hist 接口的请求日期格式，从 '2022-01-01' 为 '20220101'
 
 ## 版本更新说明
+
+1.9.3 fix: fix fx_quote_baidu interface
 
 1.9.2 fix: fix stock_xgsglb_em interface